requirements.txt
PrettyTable
colorgram.py
pandas
pyperclip
requests
python-dotenv
twilio